Membership work in progress from a while ago - not ready for use
This is some code I worked on quite a while ago, but never finished.
I hadn’t committed it because I wanted it to be working before doing so.
However, now I want to get it out of the way so it isn’t distracting from
other code I’m working on.

I do fully intend to get back to this at some point…