added is_interested field to rsvps table
A way for users to indicate their interest in an event, even if they can’t attend.
This is an attempt to get “yes” and “maybe” responses to be more accurate, since many users use them as a way to show their support even if they really plan not to attend.