WARNING: Migrations major overhaul - breaks previous migrations
You cannot migrate from previous versions to this version!

* The migrations have been significantly altered,
* previous files renumbered and renamed,
* fields arbitrarily rearranged, removed, added,
* indexes changed all over the place (some now unique),
* etc.

FUTURE CHANGES:

This will hopefully be the only time this happens.

I’m making this change now because no one else is using the code yet (that I’m aware of), so there’s still room to make this major cleanup.

The next commit will rename many of the migration files. I’m doing this in two parts so the diff on this commit will show content changes and the diff on the next commit will show renaming (doing both at once would make it much harder to figure out what all the changes are).