WARNING: Migrations renamed - breaks previous migrations
You cannot migrate from previous versions to the new migrations!

FUTURE CHANGES:

This will hopefully be the only time this happens.

I’m making this change now because no one else is using the code yet (that I’m aware of), so there’s still room to make this major cleanup.

The previous commit changed the content of the migrations quite significantly. I’ve done this in two parts so the diff on the previous commit will show content changes and the diff on this commit will show renaming (doing both at once would make it much harder to figure out what all the changes are).