tags in coded format (“domain:class=id”)
Support the specialized format of tags that allows for some structured information to be represented.

For example, Flickr supports the use of tags like “upcoming:event=123” which links a tagged photo to a specific event on the Upcoming events website. Flickr then displays a link to the event listing on Upcoming in the info for the photos. Additionally, Upcoming tracks which photos have these tags and displays them when users view the specified events.