email addresses controller - now working
The email addresses controller now works (/email_addresses or /user/email_addresses).

Made some fixes and additions to EmailAddress and Membership (including the memberships controllers and views) to get this working.

Started implementing support for sub-sections for the users controller (like with the groups controller). Still need to put content in the views/users/ partial templates for header, footer and standard.